requirements.txt
numpy>=1.26.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0